Implement binary tree data structure utility
